data-ingestion/requirements.txt
# Core Google Cloud dependencies
google-cloud-aiplatform>=1.24.0
google-cloud-storage>=2.8.0
google-cloud-bigquery>=3.0.0
google-cloud-logging>=3.0.0
google-cloud-pubsub>=2.13.0
google-api-python-client>=2.0.0

# Web framework and server
Flask>=2.0.1
gunicorn>=20.1.0
Werkzeug>=2.0.0

# Document processing
PyPDF2>=3.0.1
pdfkit>=1.0.0
weasyprint>=54.0
lxml>=4.9.0

# Utilities and helpers
requests>=2.31.0
python-dotenv>=0.19.0
pydantic>=1.8.2
typing-extensions>=4.0.0

# Development and testing
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0
mypy>=0.910

# Monitoring and logging
opencensus>=0.8.0
opencensus-ext-stackdriver>=0.8.0
prometheus-client>=0.12.0

# Security
python-jose>=3.3.0
cryptography>=3.4.0